api/pyproject.toml
[project]
name = "evolution-todo-api"
version = "0.1.0"
description = "FastAPI backend for Evolution Todo application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=5.0.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "resend>=2.0.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
